move the args parser
