+++ internal/parser/args_parser.go
@@ -1,4 +1,4 @@
-package main
+package parser
 
 import (
 	"strings"
