+++ internal/parser/args_parser_test.go
@@ -1,4 +1,4 @@
-package main
+package parser
 
 import (
 	"testing"
